BE_220930_mr : Merge branch "feat-backend/myChallenge" into "BE_develop"

See merge request s07-blockchain-contract-sub2/S07P22A708!118

+++ backend/src/main/java/com/ssafy/kirin/dto/response/ChallengeDetailDTO.java
@@ -3,5 +3,5 @@ package com.ssafy.kirin.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public record ChallengeDetailDTO(String info, List<DonationDTO> donors, String donationOraganizationName, LocalDateTime startDate, LocalDateTime endDate, Integer targetAmount, Integer targetNum, Integer currentNum, Integer currentAmount) {
+public record ChallengeDetailDTO(String info, List<DonationDTO> donors, String donationOraganizationName, LocalDateTime startDate, LocalDateTime endDate, Long targetAmount, Integer targetNum, Integer currentNum, Long currentAmount) {
 }
